version 0.0.1
sets version 0.01
fix README doi for Brieman2001
fix ull bitset on i386 build
fix build error for gcc 4.9 unordered_amp with enum key
fix sparse test bug due to incorrect seed

+++ src/globals.h
@@ -27,7 +27,7 @@ namespace literanger {
 using count_vector = std::vector<size_t>;
 using key_vector = std::vector<size_t>;
 using dbl_vector = std::vector<double>;
-using ull_bitenc = std::bitset<std::numeric_limits<size_t>::digits>;
+using ull_bitenc = std::bitset<std::numeric_limits<unsigned long long>::digits>;
 
 #ifndef M_PI
 #define M_PI 3.14159265358979323846
